all backend.test.js tests working
